refactor: avoid header access repetition
